+++ src/collective/iconifiedcategory/utils.py
@@ -268,11 +268,27 @@ def get_categorized_elements(context,
                              result_type='dict',
                              portal_type=None,
                              sort_on=None,
-                             uids=[]):
+                             uids=[],
+                             filters={}):
     """Return categorized elements.
        p_result_type may be :
        - 'dict': default, essential metadata are returned as a dict;
-       - 'objects': categorized objects are returned."""
+       - 'objects': categorized objects are returned.
+       If some p_filters are given, the values will be filtered,
+       available filters are values stored in categorized_elements."""
+    def _check_filters(infos):
+        """ """
+        keep = True
+        for k, v in filters.items():
+            # manage case when stored value is a list or not
+            stored_value = infos[k]
+            if not hasattr(stored_value, '__iter__'):
+                stored_value = (stored_value, )
+            if v not in stored_value:
+                keep = False
+                break
+        return keep
+
     elements = []
     categorized_elements = _categorized_elements(context)
     if not categorized_elements:
@@ -282,8 +298,9 @@ def get_categorized_elements(context,
     catalog = api.portal.get_tool('portal_catalog')
     current_user_allowedRolesAndUsers = catalog._listAllowedRolesAndUsers(api.user.get_current())
     for uid, infos in categorized_elements.items():
-        if uids and uid not in uids or \
-           portal_type and infos['portal_type'] != portal_type or \
+        if (uids and uid not in uids) or \
+           (portal_type and infos['portal_type'] != portal_type) or \
+           not _check_filters(infos) or \
            (infos['confidential'] and
                 not set(infos['allowedRolesAndUsers']).intersection(current_user_allowedRolesAndUsers)):
             continue