+++ CHANGES.rst
@@ -7,7 +7,7 @@ Changelog
 - Added possibility to filter displayed categorized elements on any attribute
   of the categorized element.
   [gbastien]
-- Added a ZPublisher `:json` suffix type converter.
+- Rely on imio.helpers because we need the `:json` suffix type converter.
   [gbastien]
 
 0.48 (2021-01-19)